requirements.txt
Flask==1.1.2
numpy @ file:///D:/bld/numpy_1629092219550/work
Pillow @ file:///D:/bld/pillow_1625677957583/work
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
